Split the code from AutoSchedule.cpp in 2:  * the beam search part of the code itself remains in AutoSchedule.cpp  * The generation of valide candidate to search over as well as their featurization moved to LoopNest.h
This will enable the development of new search strategies without having to duplicate all the code that now resides in LoopNest.h
